HEAVY WIP: GamepadEX
This is a pain. Why does the base gamepad class suck and why is reflection so confusing.
I will probably need to change it so that ButtonEX is two objects: "BoolButtonEX" and "DoubleButtonEX" for the two possible value types it returns for it to work properly. Or figure out reflection (unlikely).
Just needs a LOT more work still, because this does not feel computationally or developmentally efficient. Have any other libraries implemented an "Enhanced GamePad" object? Look up later.